+++ routes/api.js
@@ -25,6 +25,10 @@ const logout = require("./user/logout");
 const login = require("./user/login");
 const signup = require("./user/signup");
 const addUser = require("./user/addUser");
+const deleteUser = require("./user/deleteUser");
+const changeFname = require("./user/changeFname");
+const changeLname = require("./user/changeLname");
+const changeEmail = require("./user/changeEmail");
 
 const createClassification = require("./training/createClassification");
 const addClasses = require("./training/addClasses");
@@ -35,11 +39,13 @@ const yoloRun = require("./training/yoloRun");
 const deleteRun = require("./training/deleteRun");
 const python = require("./training/python");
 const darknet = require("./training/darknet");
-const removePath = require("./training/movePath");
+const removePath = require("./training/removePath");
 const removeDarknetPath = require("./training/removeDarknetPath");
 const removeWeights = require("./training/removeWeights");
 const removeScript = require("./training/removeScript");
 const run = require("./training/run");
+const updateClass = require("./training/updateClass");
+const deleteClass = require("./training/deleteClass");
 
 const createProject = require("./projects/createProject");
 const updateProject = require("./projects/updateProject");
@@ -56,6 +62,8 @@ const updateLabels = require("./labelling/updateLabels");
 const downloadDataset = require("./downloads/downloadDataset");
 const downloadProject = require("./downloads/downloadProject");
 const downloadScript = require("./downloads/downloadScript");
+const downloadWeights = require("./downloads/downloadWeights");
+const downloadRun = require("./downloads/downloadRun");
 
 const test = require("./tests/test");
 const mergeTest = require("./tests/mergeTest");
@@ -65,6 +73,10 @@ api.post("/logout", logout);
 api.post("/login", login);
 api.post("/signup", signup);
 api.post("/addUser", addUser);
+api.post("/deleteUser", deleteUser);
+api.post("/changeFname", changeFname);
+api.post("/changeLname", changeLname);
+api.post("/changeEmail", changeEmail);
 
 // TRAINING ROUTES
 api.post("/api/createC", createClassification);
@@ -81,6 +93,8 @@ api.post("/remove_darknet_path", removeDarknetPath);
 api.post("/remove_weights", removeWeights);
 api.post("/remove_script", removeScript);
 api.post("/run", run);
+api.post("/updateClass", updateClass);
+api.post("/deleteClass", deleteClass);
 
 // PROJECT ROUTES
 api.post("/createP", createProject);
@@ -100,707 +114,12 @@ api.post("/updateLabels", updateLabels);
 api.post("/downloadDataset", downloadDataset);
 api.post("/downloadProject", downloadProject);
 api.post("/downloadScript", downloadScript);
+api.post("/downloadWeights", downloadWeights);
+api.post("/downloadRun", downloadRun);
 
 // TEST ROUTES
 api.post("/test", test);
 api.post("/mergeTest", mergeTest);
-
-////////////////////////////Download page/////////////////////////////////////////////////
-
-api.post("/downloadWeights", async (req, res) => {
-    console.log("downloadingWeights");
-
-    // get URL variables
-    var PName = req.body.PName,
-        admin = req.body.Admin,
-        IDX = parseInt(req.body.IDX),
-        user = req.cookies.Username;
-
-    var weights_arr = [];
-    weights_arr.push(req.body["weights[]"]);
-    var weights = [];
-    weights = weights.concat.apply(weights, weights_arr).filter(Boolean);
-    console.log("weights: ", weights);
-
-    // Set paths
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + admin + "-" + PName, // $LABELING_TOOL_PATH/public/projects/project_name
-        images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/project_name/images
-        //downloads_path = project_path + '/downloads', // $LABELING_TOOL_PATH/public/projects/project_name/downloads
-        downloads_path = main_path + user + "_Downloads",
-        training_path = project_path + "/training",
-        python_path = training_path + "/python",
-        logs_path = training_path + "/logs",
-        weights_path = training_path + "/weights";
-
-    // Create zipfile
-
-    var output = fs.createWriteStream(downloads_path + "/weights.zip");
-    var archive = archiver("zip");
-
-    output.on("close", function () {
-        console.log(archive.pointer() + " total bytes");
-        console.log(
-            "archiver has been finalized and the output file descriptor has closed.",
-        );
-        res.download(downloads_path + "/weights.zip");
-    });
-    archive.on("error", function (err) {
-        throw err;
-    });
-
-    archive.pipe(output);
-    for (var i = 0; i < weights.length; i++) {
-        var weights_file = weights_path + "/" + weights[i];
-        archive.file(weights_file, { name: weights[i] });
-    }
-    archive.finalize();
-});
-
-api.post("/downloadRun", async (req, res) => {
-    console.log("downloadRun");
-
-    var PName = req.body.PName,
-        Admin = req.body.Admin,
-        IDX = req.body.IDX,
-        weights = req.body.weights,
-        user = req.cookies.Username,
-        log_file = req.body.log_file,
-        run_path = req.body.run_path;
-
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + user + "-" + PName, // $LABELING_TOOL_PATH/public/projects/project_name
-        images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/project_name/images
-        downloads_path = main_path + user + "_Downloads",
-        training_path = project_path + "/training",
-        logs_path = training_path + "/logs/";
-
-    var output = fs.createWriteStream(
-        downloads_path + "/" + log_file.substr(0, log_file.length - 4) + ".zip",
-    );
-    var archive = archiver("zip");
-
-    output.on("close", function () {
-        console.log(archive.pointer() + " total bytes");
-        console.log(
-            "archiver has been finalized and the output file descriptor has closed.",
-        );
-        res.download(
-            downloads_path +
-                "/" +
-                log_file.substr(0, log_file.length - 4) +
-                ".zip",
-        );
-    });
-    archive.on("error", function (err) {
-        throw err;
-    });
-
-    archive.pipe(output);
-
-    var logs = await readdirAsync(`${run_path}`);
-
-    for (var i = 0; i < logs.length; i++) {
-        archive.file(`${run_path}${logs[i]}`, { name: logs[i] });
-    }
-
-    archive.finalize();
-});
-
-//TODO
-//Call Ashwin's script to configure YOLO files
-api.post("/yolo-run", async (req, res) => {
-    console.log("yolo-run");
-
-    const { exec } = require("child_process");
-    //const {spawn} = require('child_process');
-
-    var date = Date.now();
-
-    var PName = req.body.PName,
-        Admin = req.body.Admin,
-        user = req.cookies.Username,
-        darknet_path = req.body.darknet_path,
-        log = `${date}.log`,
-        trainDataPer = req.body.TrainingPercent,
-        batch = req.body.batch,
-        subdiv = req.body.subdiv,
-        width = req.body.width,
-        height = req.body.height,
-        weight_name = req.body.weights;
-
-    var err_file = `${date}-error.log`;
-
-    /*
-	Steps:
-		1. Create txt files for each image
-		2. Create classes.txt file
-		3. Call datatovalues.py script
-	*/
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/", // $LABELING_TOOL_PATH/public/projects/
-        project_path = main_path + Admin + "-" + PName, // $LABELING_TOOL_PATH/public/projects/project_name
-        images_path = project_path + "/images", // $LABELING_TOOL_PATH/public/projects/project_name/images
-        downloads_path = main_path + user + "_Downloads",
-        training_path = project_path + "/training",
-        logs_path = training_path + "/logs",
-        run_path = `${logs_path}/${date}`,
-        classes_path = run_path + "/classes.txt",
-        weight_path = training_path + "/weights/" + weight_name,
-        yolo_script = public_path + "controllers/training/datatovalues.py",
-        wrapper_path =
-            public_path + "controllers/training/train_data_from_project.py";
-
-    //Create run path
-    if (!fs.existsSync(run_path)) {
-        fs.mkdirSync(run_path);
-    }
-
-    fs.writeFile(`${run_path}/${log}`, "", (err) => {
-        if (err) throw err;
-    });
-
-    //////////////////////Copy YOLO template files over to run folder///////////////////////////////////////////////////
-    console.log("Copy YOLO template files over to run folder");
-    cfgTemp_path = public_path + "controllers/training/cfgTemplate.txt";
-    cfgTemp = run_path + "/cfgTemplate.txt";
-    fs.copyFile(cfgTemp_path, cfgTemp, (err) => {
-        if (err) {
-            console.log(err);
-        }
-    });
-
-    dataTemp_path = public_path + "controllers/training/dataTemplate.txt";
-    dataTemp = run_path + "/dataTemplate.txt";
-    fs.copyFile(dataTemp_path, dataTemp, (err) => {
-        if (err) {
-            console.log(err);
-        }
-    });
-
-    /////////////////////Copy darknet config script to darknet directory///////////////////////////////////////////
-    console.log("Copy darknet config script to darknet directory");
-    darknet_cfg_script = darknet_path + "/datatovalues.py";
-    if (!existsSync(darknet_cfg_script)) {
-        fs.copyFile(yolo_script, darknet_cfg_script, (err) => {
-            if (err) {
-                console.log(err);
-            }
-        });
-    }
-
-    /////////////////////////Create Project within darknet if does not exist///////////////////////////////////////
-    console.log("Create Project within darknet if does not exist");
-    project = `${Admin}-${PName}`;
-    abs_darknet_project_path = path.join(darknet_path, project);
-    abs_darknet_images_path = path.join(abs_darknet_project_path, "images");
-
-    if (!fs.existsSync(abs_darknet_images_path)) {
-        fs.mkdirSync(abs_darknet_project_path);
-        //Create symbolic link to images
-        fs.symlink(images_path, abs_darknet_images_path, "dir", (err) => {
-            if (err) {
-                console.log(err);
-            } else {
-                console.log("Images Symlink created");
-                console.log(
-                    "Symlink is a directory: ",
-                    fs.lstatSync(abs_darknet_images_path).isDirectory(),
-                );
-            }
-        });
-    }
-
-    //////////////////////////Create symbolic link from darknet to run/////////////////////////////////////////////
-    console.log("Create symbolic link from darknet to run");
-    abs_darknet_project_run = path.join(
-        abs_darknet_project_path,
-        date.toString(),
-    );
-
-    if (!fs.existsSync(abs_darknet_project_run)) {
-        fs.symlink(run_path, abs_darknet_project_run, "dir", (err) => {
-            if (err) {
-                console.log(err);
-            } else {
-                console.log("Symlink created");
-                console.log(
-                    "Symlink is a directory: ",
-                    fs.lstatSync(abs_darknet_project_run).isDirectory(),
-                );
-            }
-        });
-    }
-
-    // Connect to database
-    var ycdb = new sqlite3.Database(
-        project_path + "/" + PName + ".db",
-        (err) => {
-            if (err) {
-                return console.error(err.message);
-            }
-            console.log("Connected to ycdb.");
-        },
-    );
-    ycdb.getAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.get(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-        });
-    };
-    ycdb.allAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.all(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-        });
-    };
-    ycdb.runAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.run(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-        });
-    };
-
-    /////////////////////////////////////////Create label txt files///////////////////////////////////////////////////
-    console.log("Create label txt files");
-    var cnames = [];
-    var results1 = await ycdb.allAsync("SELECT * FROM Classes");
-    var results2 = await ycdb.allAsync("SELECT * FROM Images");
-    for (var i = 0; i < results1.length; i++) {
-        cnames.push(results1[i].CName);
-    }
-    var dict_images_labels = {};
-
-    for (var i = 0; i < results2.length; i++) {
-        var img = fs.readFileSync(`${images_path}/${results2[i].IName}`),
-            img_data = probe.sync(img),
-            img_w = img_data.width,
-            img_h = img_data.height;
-
-        var results3 = await ycdb.allAsync(
-            "SELECT * FROM Labels WHERE IName = '" + results2[i].IName + "'",
-        );
-        for (var j = 0; j < results3.length; j++) {
-            // x, y, w, h
-            var centerX = (results3[j].X + results3[j].W / 2) / img_w;
-            var centerY = (results3[j].Y + results3[j].H / 2) / img_h;
-            to_string_value =
-                cnames.indexOf(results3[j].CName) +
-                " " +
-                centerX +
-                " " +
-                centerY +
-                " " +
-                results3[j].W / img_w +
-                " " +
-                results3[j].H / img_h +
-                "\n";
-            if (dict_images_labels[results2[i].IName] == undefined) {
-                dict_images_labels[results2[i].IName] = to_string_value;
-            } else {
-                dict_images_labels[results2[i].IName] += to_string_value;
-            }
-        }
-        if (results3.length == 0) {
-            dict_images_labels[results2[i].IName] = "";
-        }
-    }
-
-    for (var key in dict_images_labels) {
-        remove_dot_ext = key.split(".")[0];
-        fs.writeFileSync(
-            `${images_path}/${remove_dot_ext}.txt`,
-            dict_images_labels[key],
-            (err) => {
-                if (err) throw err;
-            },
-        );
-    }
-
-    //////////////////////////////////////////////////Create Classes file///////////////////////////////////////////////////////////////////
-    console.log("Create Classes file");
-    var classes = "";
-    for (var i = 0; i < results1.length; i++) {
-        classes = classes + results1[i].CName + "\n";
-    }
-    classes = classes.substring(0, classes.length - 1); //remove trailing newline
-
-    fs.writeFileSync(classes_path, classes, (err) => {
-        if (err) throw err;
-        console.log("done writing Classes file");
-    });
-
-    ycdb.close((err) => {
-        if (err) {
-            console.log(err);
-        } else {
-            console.log("ycdb closed successfully");
-        }
-    });
-
-    //////////////////////////////////////////// Call Ashwin's script here//////////////////////////////////////////////////////////////////
-    console.log("Calling Darknet");
-    // Pass in python path, script, and options
-    darknet_project_run = path.join(project, date.toString());
-    darknet_images_path = path.join(abs_darknet_project_path, "images");
-    var cmd = `python3 ${yolo_script} -d ${darknet_path}/${darknet_project_run} -i ${darknet_images_path} -n ${classes_path} -p ${trainDataPer} -l ${abs_darknet_project_run}/${log} -y ${darknet_path} -w ${weight_path} -b ${batch} -s ${subdiv} -x ${width} -t ${height}`;
-    var success = "";
-    var error = "";
-    console.log(cmd);
-
-    //change directory to darknet
-    process.chdir(darknet_path);
-
-    var child = exec(cmd, (err, stdout, stderr) => {
-        if (err) {
-            console.log(`This is the error: ${err.message}`);
-            if (err.message != "stdout maxBuffer length exceeded") {
-                success = err.message;
-                fs.writeFile(
-                    `${darknet_project_run}/${err_file}`,
-                    success,
-                    (err) => {
-                        if (err) throw err;
-                    },
-                );
-            }
-        } else if (stderr) {
-            console.log(`This is the stderr: ${stderr}`);
-            if (stderr != "stdout maxBuffer length exceeded") {
-                fs.writeFile(
-                    `${darknet_project_run}/${err_file}`,
-                    stderr,
-                    (err) => {
-                        if (err) throw err;
-                    },
-                );
-            }
-            //return;
-        }
-        console.log("stdout: ", stdout);
-        console.log("stderr: ", stderr);
-        console.log("err: ", err);
-        console.log("The script has finished running");
-        fs.writeFile(`${darknet_project_run}/done.log`, success, (err) => {
-            if (err) throw err;
-        });
-    });
-    res.send({ Success: `Training Started` });
-});
-
-module.exports = api;
-
-api.post("/deleteUser", async (req, res) => {
-    console.log("delete user");
-    var user = req.cookies.Username;
-
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/";
-
-    var darknet_path = new Set();
-
-    var filesinfolder = readdirSync(main_path);
-    // Delete the Files
-    for (var i = 0; i < filesinfolder.length; i++) {
-        if (
-            filesinfolder[i].split("_")[0] == user ||
-            filesinfolder[i].split("-")[0] == user
-        ) {
-            if (filesinfolder[i].split("-")[0] == user) {
-                var tempdarknet = fs
-                    .readFileSync(
-                        main_path +
-                            filesinfolder[i] +
-                            "/training/darknetPaths.txt",
-                        "utf-8",
-                    )
-                    .split("\n")
-                    .join(",")
-                    .split(",");
-                for (var f = 0; f < tempdarknet.length; f++) {
-                    if (tempdarknet[f] != "") {
-                        darknet_path.add(tempdarknet[f]);
-                    }
-                }
-            }
-            rimraf(main_path + filesinfolder[i], (err) => {
-                if (err) {
-                    console.error(
-                        "there was an error with the user contents: ",
-                        err,
-                    );
-                } else {
-                    console.log("User directory contents successfuly deleted");
-                }
-            });
-        }
-    }
-
-    //Delete the YOLO Files
-    const drknt_temp = darknet_path.values();
-    for (var i = 0; i < darknet_path.size; i++) {
-        var current_darknet_path = drknt_temp.next().value;
-        var darknetFiles = readdirSync(current_darknet_path);
-        for (var i = 0; i < darknetFiles.length; i++) {
-            if (darknetFiles[i].split("-")[0] == user) {
-                rimraf(current_darknet_path + "/" + darknetFiles[i], (err) => {
-                    if (err) {
-                        console.error(
-                            "there was an error with the user contents: ",
-                            err,
-                        );
-                    } else {
-                        console.log(
-                            "Darknet user project contents successfuly deleted",
-                        );
-                    }
-                });
-            }
-        }
-    }
-    // Delete from the Database
-    db.getAsync("DELETE FROM Access WHERE Admin = '" + user + "'");
-    db.getAsync("DELETE FROM Access WHERE Username = '" + user + "'");
-    db.getAsync("DELETE FROM Projects WHERE Admin = '" + user + "'");
-    db.getAsync("DELETE FROM Users WHERE Username = '" + user + "'");
-
-    res.clearCookie("Username");
-    return res.send({ Success: "Yes" });
-});
-
-api.post("/updateClass", async (req, res) => {
-    console.log("Request body:", req.body);
-
-    const className = req.body.currentClassName;
-    const updateClassName = req.body.updatedValue;
-    //update Class name
-
-    var IDX = parseInt(req.body.IDX),
-        PName = req.body.PName,
-        admin = req.body.Admin,
-        user = req.cookies.Username;
-
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/",
-        project_path = main_path + user + "-" + PName;
-
-    // console.log("project_path: ", project_path);
-    // console.log("Current Class Name:", className);
-    // console.log("Updated Class Name:", updateClassName);
-
-    var db = new sqlite3.Database(project_path + "/" + PName + ".db", (err) => {
-        if (err) {
-            return console.error(err.message);
-        }
-        console.log("Connected to db.");
-    });
-
-    const updateLabels = `UPDATE Labels SET CName = ? WHERE CName = ?`;
-    const updateValidation = `UPDATE Validation SET CName = ? WHERE CName = ?`;
-    const updateClasses = `UPDATE Classes SET CName = ? WHERE CName = ?`;
-
-    db.run(updateLabels, [updateClassName, className], function (err) {
-        if (err) {
-            console.error("Error updating Labels:", err.message);
-            return res.status(500).send("Error updating Labels");
-        }
-        db.run(updateValidation, [updateClassName, className], function (err) {
-            if (err) {
-                console.error("Error updating Validation:", err.message);
-                return res.status(500).send("Error updating Validation");
-            }
-            db.run(updateClasses, [updateClassName, className], function (err) {
-                if (err) {
-                    console.error("Error updating Classes:", err.message);
-                    return res.status(500).send("Error updating Classes");
-                }
-                res.status(200).send("Class name updated successfully.");
-            });
-        });
-    });
-});
-
-api.post("/deleteClass", async (req, res) => {
-    console.log("body", req.body);
-
-    var IDX = parseInt(req.body.IDX),
-        PName = req.body.PName,
-        admin = req.body.Admin,
-        user = req.cookies.Username,
-        classes = req.body["classArray[]"];
-
-    console.log("classes: ", classes);
-    // set paths
-    var public_path = __dirname.replace("routes", ""),
-        main_path = public_path + "public/projects/",
-        project_path = main_path + admin + "-" + PName;
-
-    var dcdb = new sqlite3.Database(
-        project_path + "/" + PName + ".db",
-        (err) => {
-            if (err) {
-                return console.error(err.message);
-            }
-            console.log("Connected to dcdb.");
-        },
-    );
-    dcdb.getAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.get(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-        });
-    };
-    dcdb.allAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.all(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-        });
-    };
-    dcdb.runAsync = function (sql) {
-        var that = this;
-        return new Promise(function (resolve, reject) {
-            that.run(sql, function (err, row) {
-                if (err) {
-                    console.log("runAsync ERROR! ", err);
-                    reject(err);
-                } else resolve(row);
-            });
-        }).catch((err) => {
-            console.error(err);
-        });
-    };
-
-    if (classes.includes(",")) {
-        classes = classes.split(",");
-    }
-    console.log(classes);
-
-    //Delete classes
-    var deleteLabels = "";
-    var deleteClasses = "";
-    var deleteValid = "";
-    if (typeof classes == "string") {
-        deleteClasses = `DELETE FROM Classes WHERE CName = '${classes}'`;
-        deleteLabels = `DELETE FROM Labels WHERE CName = '${classes}'`;
-        deleteValid = `DELETE FROM Validation WHERE CName = '${classes}'`;
-    } else {
-        deleteLabels = `DELETE FROM Labels WHERE CName = '${classes[0]}'`;
-        deleteClasses = `DELETE FROM Classes WHERE CName = '${classes[0]}'`;
-        deleteValid = `DELETE FROM Validation WHERE CName = '${classes[0]}'`;
-
-        for (var i = 1; i < classes.length; i++) {
-            var string = ` OR CName = '${classes[i]}'`;
-            deleteLabels += string;
-            deleteClasses += string;
-            deleteValid += string;
-        }
-    }
-    console.log(deleteClasses);
-    await dcdb.runAsync(deleteClasses);
-
-    console.log(deleteLabels);
-    await dcdb.runAsync(deleteLabels);
-
-    console.log(deleteValid);
-    await dcdb.runAsync(deleteValid);
-
-    console.log("deleteClass (redirect)");
-
-    dcdb.close(function (err) {
-        if (err) {
-            console.error(err);
-        } else {
-            console.log("cdb closed successfully");
-        }
-    });
-
-    return res.redirect("/config?IDX=" + IDX);
-});
-
-api.post("/changeFname", async (req, res) => {
-    console.log("changeFname");
-
-    var user = req.cookies.Username,
-        FName = req.body.FName;
-    await db.runAsync(
-        "UPDATE Users SET FirstName = '" +
-            FName +
-            "' WHERE Username = '" +
-            user +
-            "'",
-    );
-
-    return res.send({ Success: "Yes" });
-});
-
-api.post("/changeLname", async (req, res) => {
-    console.log("changeLname");
-
-    var user = req.cookies.Username,
-        LName = req.body.LName;
-    await db.runAsync(
-        "UPDATE Users SET LastName = '" +
-            LName +
-            "' WHERE Username = '" +
-            user +
-            "'",
-    );
-
-    return res.send({ Success: "Yes" });
-});
-
-api.post("/changeEmail", async (req, res) => {
-    console.log("changeEmail");
-
-    var user = req.cookies.Username,
-        Email = req.body.Email;
-    await db.runAsync(
-        "UPDATE Users SET Email = '" +
-            Email +
-            "' WHERE Username = '" +
-            user +
-            "'",
-    );
-
-    return res.send({ Success: "Yes" });
-});
 
 api.post("/changeUname", async (req, res) => {
     console.log("changeUname");
@@ -1833,3 +1152,5 @@ api.post("/solo-change-class", async (req, res) => {
     );
     res.send({ Success: "Yes" });
 });
+
+module.exports = api;
